CHORE: improve file & folder structure
