+++ src/components/Projects/index.tsx
@@ -1,0 +1,5 @@
+import React from "react";
+
+export default function Projects() {
+  return <div>Projects</div>;
+}
